app/requirements.txt
fastapi==0.115.0
starlette==0.38.6
pydantic==2.9.2
typing-extensions==4.12.2
fastapi-cli==0.0.5
httpx==0.27.2
jinja2==3.1.4
python-multipart==0.0.12
email-validator==2.2.0
uvicorn==0.31.1
dnspython==2.7.0
idna==3.10
typer==0.12.5
anyio==4.6.0
certifi==2024.8.30
httpcore==1.0.6
sniffio==1.3.1
h11==0.14.0
MarkupSafe==3.0.1
annotated-types==0.7.0
pydantic-core==2.23.4
click==8.1.7
colorama==0.4.6
httptools==0.6.1
python-dotenv==1.0.1
pyyaml==6.0.2
watchfiles==0.24.0
websockets==13.1
shellingham==1.5.4
rich==13.9.2
markdown-it-py==3.0.0
pygments==2.18.0
mdurl==0.1.2
sqlmodel==0.0.22
SQLAlchemy==2.0.35
greenlet==3.1.1
psycopg2==2.9.9
jose==1.0.0
cffi==1.17.1
cryptography==43.0.1
ecdsa==0.19.0
pyasn1==0.6.1
pycparser==2.22
python-jose==3.3.0
rsa==4.9
six==1.16.0
bcrypt==4.2.0
passlib==1.7.4
Faker==30.3.0
python-dateutil==2.9.0.post0
Werkzeug==3.0.4
Pillow==10.4.0